move leaderboard calculation into own concern
