+++ app/models/user_leaderboard.rb
@@ -1,6 +1,7 @@
 class UserLeaderboard < ActiveRecord::Base
   include CurrentWeek
-
+  include LeaderboardCalculation
+  
   belongs_to :user
 
   def name
