+++ spec/models/user_spec.rb
@@ -44,32 +44,6 @@ describe User do
     end
   end
 
-  describe ".calculate_leaderboard" do
-    let!(:user) { create :user }
-
-    context 'without entries' do
-      it 'is not generating the leaderboard' do
-        expect do
-          User.calculate_leaderboard
-        end.to change(UserLeaderboard.where(total_minutes: 0), :count).by(1)
-      end
-    end
-
-    context 'with entries' do
-      let!(:entry) { create :entry, user: user }
-
-      it 'is generating the leaderboard' do
-        expect { User.calculate_leaderboard }.to change(
-          UserLeaderboard.where(total_minutes: 0),
-          :count
-        ).by(0)
-
-        expect(UserLeaderboard.where(
-          total_minutes: entry.minutes).count).to eq(1)
-      end
-    end
-  end
-
   describe '#minutes_of_current_week' do
     let(:user)   { create(:user) }
     let!(:entry) { create(:entry, minutes: 1337, user_id: user.id) }
